pyproject.toml
[project]
name = "otus_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.12",
    "gunicorn>=23.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "sqladmin[full]>=0.20.1",
    "sqlalchemy[asyncio]>=2.0.40",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
